# api/swagger/gateway/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Платежный шлюз для E-commerce системы",
        "contact": {
            "name": "Курмыза Павел",
            "email": "tmrrwnxtsn@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/tool": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Платежные средства"
                ],
                "summary": "Получить список сохраненных платежных средств клиента",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор клиента",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Идентификатор сессии клиента",
                        "name": "session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Код языка, обозначение по RFC 5646",
                        "name": "lang_code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/v1.toolListResponse"
                        }
                    },
                    "default": {
                        "description": "Ответ с ошибкой",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    }
                }
            }
        },
        "/tool/edit": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Платежные средства"
                ],
                "summary": "Изменить информацию о платежном средстве",
                "parameters": [
                    {
                        "description": "Тело запроса",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.toolEditRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/v1.toolEditResponse"
                        }
                    },
                    "default": {
                        "description": "Ответ с ошибкой",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    }
                }
            }
        },
        "/tool/remove": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Платежные средства"
                ],
                "summary": "Удалить платежное средство",
                "parameters": [
                    {
                        "description": "Тело запроса",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.toolRemoveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/v1.toolRemoveResponse"
                        }
                    },
                    "default": {
                        "description": "Ответ с ошибкой",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "v1.errorContent": {
            "type": "object",
            "required": [
                "code",
                "description"
            ],
            "properties": {
                "code": {
                    "description": "Код ошибки",
                    "type": "string",
                    "example": "InvalidRequest"
                },
                "description": {
                    "description": "Описание ошибки для разработки",
                    "type": "string",
                    "example": "user_id param is required"
                },
                "message": {
                    "description": "Сообщение об ошибке для клиента",
                    "type": "string",
                    "example": "Internal server error occurred. Please try again later."
                }
            }
        },
        "v1.errorResponse": {
            "type": "object",
            "required": [
                "error",
                "success"
            ],
            "properties": {
                "error": {
                    "description": "Развернутая информация об ошибке",
                    "allOf": [
                        {
                            "$ref": "#/definitions/v1.errorContent"
                        }
                    ]
                },
                "success": {
                    "description": "Результат обработки запроса (всегда false)",
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "v1.tool": {
            "type": "object",
            "required": [
                "caption",
                "external_method",
                "id",
                "name",
                "type"
            ],
            "properties": {
                "caption": {
                    "description": "Значение платежного средства, например:\n* Маскированная банковская карта\n* Номер электронного кошелька\n* Адрес электронной почты\n* и т.д.",
                    "type": "string",
                    "example": "444444******4444"
                },
                "details": {
                    "description": "Дополнительная информация о платежном средстве",
                    "allOf": [
                        {
                            "$ref": "#/definitions/v1.toolDetails"
                        }
                    ]
                },
                "external_method": {
                    "description": "Внутренний код платежного метода платежной системы, к которой относится платежное средство",
                    "type": "string",
                    "example": "yookassa_bank_card"
                },
                "id": {
                    "description": "Идентификатор платежного средства",
                    "type": "string",
                    "example": "2dc32aa0-000f-5000-8000-16d7bc6cd09f"
                },
                "name": {
                    "description": "Название платежного средства",
                    "type": "string",
                    "example": "Карта брата"
                },
                "type": {
                    "description": "Тип платежного средства:\n* Банковская карта - \"card\"\n* Электронный кошелек - \"wallet\"",
                    "type": "string",
                    "example": "card"
                }
            }
        },
        "v1.toolDetails": {
            "type": "object",
            "properties": {
                "bank_name": {
                    "description": "Название банка, выпустившего банковскую карту",
                    "type": "string",
                    "example": "Sberbank"
                },
                "card_holder": {
                    "description": "Владелец банковской карты",
                    "type": "string",
                    "example": "Ivanov Ivan"
                },
                "card_type": {
                    "description": "Тип банковской карты",
                    "type": "string",
                    "example": "Visa"
                },
                "expiry_month": {
                    "description": "Срок действия банковской карты (месяц, MM)",
                    "type": "integer",
                    "example": 10
                },
                "expiry_year": {
                    "description": "Срок действия банковской карты (год, YYYY)",
                    "type": "integer",
                    "example": 2023
                },
                "wallet_number": {
                    "description": "Номер электронного кошелька",
                    "type": "string",
                    "example": "410011758831136"
                }
            }
        },
        "v1.toolEditRequest": {
            "type": "object",
            "required": [
                "external_method",
                "id",
                "lang_code",
                "name",
                "session_id",
                "user_id"
            ],
            "properties": {
                "external_method": {
                    "description": "Внутренний код платежного метода платежной системы, к которой относится платежное средство",
                    "type": "string",
                    "example": "yookassa_bank_card"
                },
                "id": {
                    "description": "Идентификатор платежного средства",
                    "type": "string",
                    "example": "2dc32aa0-000f-5000-8000-16d7bc6cd09f"
                },
                "lang_code": {
                    "description": "Код языка, обозначение по RFC 5646",
                    "type": "string",
                    "example": "en"
                },
                "name": {
                    "description": "Новое название платежного инструмента, выбранное клиентом",
                    "type": "string",
                    "example": "Карта брата"
                },
                "session_id": {
                    "description": "Идентификатор сессии клиента",
                    "type": "string",
                    "example": "LRXZmXPGusPCfys48LadjFew"
                },
                "user_id": {
                    "description": "Идентификатор клиента",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "v1.toolEditResponse": {
            "type": "object",
            "required": [
                "success",
                "tool"
            ],
            "properties": {
                "success": {
                    "description": "Результат обработки запроса (всегда true)",
                    "type": "boolean",
                    "example": true
                },
                "tool": {
                    "description": "Информация об измененном платежном средстве",
                    "allOf": [
                        {
                            "$ref": "#/definitions/v1.tool"
                        }
                    ]
                }
            }
        },
        "v1.toolListResponse": {
            "type": "object",
            "required": [
                "success",
                "tools"
            ],
            "properties": {
                "success": {
                    "description": "Результат обработки запроса (всегда true)",
                    "type": "boolean",
                    "example": true
                },
                "tools": {
                    "description": "Массив сохраненных платежных средств клиента",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.tool"
                    }
                }
            }
        },
        "v1.toolRemoveRequest": {
            "type": "object",
            "required": [
                "external_method",
                "id",
                "lang_code",
                "session_id",
                "user_id"
            ],
            "properties": {
                "external_method": {
                    "description": "Внутренний код платежного метода платежной системы, к которой относится платежное средство",
                    "type": "string",
                    "example": "yookassa_bank_card"
                },
                "id": {
                    "description": "Идентификатор платежного средства",
                    "type": "string",
                    "example": "2dc32aa0-000f-5000-8000-16d7bc6cd09f"
                },
                "lang_code": {
                    "description": "Код языка, обозначение по RFC 5646",
                    "type": "string",
                    "example": "en"
                },
                "session_id": {
                    "description": "Идентификатор сессии клиента",
                    "type": "string",
                    "example": "LRXZmXPGusPCfys48LadjFew"
                },
                "user_id": {
                    "description": "Идентификатор клиента",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "v1.toolRemoveResponse": {
            "type": "object",
            "required": [
                "message",
                "success"
            ],
            "properties": {
                "message": {
                    "description": "Сообщение, которое необходимо показать клиенту",
                    "type": "string",
                    "example": "Платежное средство удалено."
                },
                "success": {
                    "description": "Результат обработки запроса (всегда true)",
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Секретный ключ",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}